fix a possible issue with double coordinates

+++ framework/gwt/src/cc/alcina/framework/gwt/client/util/RelativePopupPositioning.java
@@ -150,7 +150,7 @@ public class RelativePopupPositioning {
 				rpp.getElement()
 						.getStyle()
 						.setLeft(
-								Integer.parseInt(pxl.replace("px", ""))
+								Double.parseDouble(pxl.replace("px", ""))
 										- (rl + rw - x - bwW), Unit.PX);
 			}
 		}
